feat: add GitHub download type support
Introduced a new enum, GitHubDownloadType, to specify the method for
downloading files from GitHub. This allows for flexibility in using
either the native fetch API or the Octokit Core library for
authenticated requests. Updated relevant functions and tests to
accommodate this new functionality.

BREAKING CHANGE: The signature of downloadGlobalTypesAsync has
changed to include the new githubDownloadType parameter.